use combine reducer for data and game part of state
