+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./components/App/App";
 import {applyMiddleware, compose, createStore} from "redux";
-import {reducer} from "./reducer";
+import {rootReducer as reducer} from "./reducer";
 import {Provider} from "react-redux";
 import thunk from "redux-thunk";
 import api from "./api";
